+++ cmd/hotplex/main.go
@@ -13,7 +13,7 @@ import (
 )
 
 var (
-	version   = "v1.40.0"
+	version   = "v1.41.0"
 	buildTime = "unknown"
 )